chore: rename router file to be router-builder
